pyproject.toml
[project]
name = "g70k-trae"
version = "0.1.0"
description = "Sistema de Garantia Viemar - Gestão de garantias de veículos"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "beautifulsoup4>=4.13.5",
    "email-validator>=2.3.0",
    "httpx>=0.28.1",
    "monsterui>=1.0.26",
    "playwright>=1.55.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.3.0",
    "pytest-playwright>=0.7.1",
    "python-fasthtml>=0.12.25",
    "requests>=2.32.5",
    "werkzeug>=3.1.3",
]

[tool.pytest.ini_options]
# Configuração do pytest para o projeto Viemar Garantia
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Marcadores personalizados
markers = [
    "slow: marca testes que são lentos para executar",
    "integration: marca testes de integração",
    "unit: marca testes unitários",
    "auth: marca testes relacionados à autenticação",
    "database: marca testes relacionados ao banco de dados",
    "routes: marca testes de rotas",
    "models: marca testes de modelos",
    "email: marca testes relacionados ao envio de emails",
    "e2e: marca testes end-to-end",
    "playwright: marca testes que usam Playwright"
]

# Opções padrão
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=app",
    "--cov=models",
    "--cov-report=html:htmlcov",
    "--cov-report=term-missing",
    "--cov-report=xml",
    "--cov-fail-under=85"
]

# Filtros de warnings
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]

# Configurações de logging durante os testes
log_cli = true
log_cli_level = "INFO"
